# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "AGENT STATS",
    "description": "AGENT STATS STATUSES AND CHANNEL DATA"
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    },
    {
      "url": "https://localhost:3000/"
    }
  ],
...
  "paths": {
...
    "/user/": {
      "post": {
        "tags": [],
        "description": "create new out dial jobs.",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/CreateUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "get": {
        "tags": [],
        "description": "all sip peers.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/user/{id}": {
      "put": {
        "tags": [],
        "description": "update outbound job by id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/UpdateJob"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "delete": {
        "tags": [],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      },
      "get": {
        "tags": [],
        "description": "all sip peers.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "all sip peers.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              },
              "application/xml": {
                "schema": {
                  "$ref": "#/components/schemas/SipPeers"
                }
              }
            }
          }
        },
        "security": [
          {
            "apiKeyAuth": []
          }
        ]
      }
    },
    "/user/token": {
      "post": {
        "tags": [
          "user"
        ],
        "description": "Endpoint to sign in a specific user",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AddUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/sign-in": {
      "post": {
        "tags": [
          "user"
        ],
        "description": "Endpoint to sign in a specific user",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AddUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/app/sign-in": {
      "post": {
        "tags": [
          "user"
        ],
        "description": "Endpoint to sign in a specific user",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AddUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/user/dashboard/stats": {
      "get": {
        "tags": [
          "user"
        ],
        "description": "Endpoint to sign in a specific user",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "User information.",
            "required": true,
            "schema": {
              "$ref": "#/components/schemas/AddUser"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateUser": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "test@test.com"
          },
          "password": {
            "type": "string",
            "example": "123321"
          }
        },
        "xml": {
          "name": "CreateUser"
        }
      },
...
      "UpdateJob": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "test"
          },
          "type": {
            "type": "string",
            "example": "predictive"
          },
          "pause": {
            "type": "number",
            "example": 0
          },
          "complete": {
            "type": "number",
            "example": 0
          },
          "startTime": {
            "type": "string",
            "example": "2022-08-10 20:00:00"
          },
          "endTime": {
            "type": "string",
            "example": "2022-08-10 23:00:00"
          },
          "jobData": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "number": {
                  "type": "string",
                  "example": "231203"
                },
                "mask": {
                  "type": "string",
                  "example": "123456"
                },
                "context": {
                  "type": "string",
                  "example": "default"
                },
                "trunk": {
                  "type": "string",
                  "example": "outgoing"
                }
              }
            }
          }
        },
        "xml": {
          "name": "UpdateJob"
        }
      },
...
      "SipPeers": {
        "type": "object",
        "properties": {
          "response": {
            "type": "object",
            "properties": {
              "success": {
                "type": "boolean",
                "example": true
              },
              "data": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "event": {
                      "type": "string",
                      "example": "PeerEntry"
                    },
                    "actionid": {
                      "type": "string",
                      "example": "1660286907216"
                    },
                    "channeltype": {
                      "type": "string",
                      "example": "SIP"
                    },
                    "objectname": {
                      "type": "string",
                      "example": "salman"
                    },
                    "chanobjecttype": {
                      "type": "string",
                      "example": "peer"
                    },
                    "ipaddress": {
                      "type": "string",
                      "example": "-none-"
                    },
                    "ipport": {
                      "type": "string",
                      "example": "0"
                    },
                    "dynamic": {
                      "type": "string",
                      "example": "yes"
                    },
                    "autoforcerport": {
                      "type": "string",
                      "example": "no"
                    },
                    "forcerport": {
                      "type": "string",
                      "example": "no"
                    },
                    "autocomedia": {
                      "type": "string",
                      "example": "no"
                    },
                    "comedia": {
                      "type": "string",
                      "example": "no"
                    },
                    "videosupport": {
                      "type": "string",
                      "example": "no"
                    },
                    "textsupport": {
                      "type": "string",
                      "example": "no"
                    },
                    "acl": {
                      "type": "string",
                      "example": "no"
                    },
                    "status": {
                      "type": "string",
                      "example": "UNKNOWN"
                    },
                    "realtimedevice": {
                      "type": "string",
                      "example": "no"
                    },
                    "description": {
                      "type": "string",
                      "example": ""
                    },
                    "accountcode": {
                      "type": "string",
                      "example": ""
                    }
                  }
                }
              }
            }
          }
        },
        "xml": {
          "name": "SipPeers"
        }
      },
...
      "AddUser": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "BEN@gmail.com"
          },
          "password": {
            "type": "string",
            "example": "XYZ"
          }
        },
        "required": [
          "email",
          "password"
        ],
        "xml": {
          "name": "AddUser"
        }
      },
...
    },
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "Authorization",
        "description": "any description..."
      }
    }
  }
}